2.8.6 Added voice commands, maps, and settings window

+++ app/src/main/java/com/example/my_opencv/DroneListener.java
@@ -10,7 +10,7 @@ public interface DroneListener {
 
     void onOnlineStatus(boolean online);
 
-    void onSetAppData(int [] data);
+    void onSetAppData(double [] data);
 
     void onGetAppData();
